Remove safeTypeCast
Summary:
From what I can tell, type punning through a union is still technically
UB in C++. In many of the cases we use it, a `reinterpret_cast` is
sufficient, since it is explicitly allowed to convert pointers to
integers and vice versa.

For doing the double to uint64_t conversion, LLVM provides utility
functions that internally use memcpy.

Reviewed By: jpporto

Differential Revision: D35135347

fbshipit-source-id: d7d1dba903e69aaea8b7d174a8737a8d6e287fcc