Improve comment attachment logic
Summary:
Currently we have no way of specifying comment placement outside of "leading" or "trailing" and whether you get an inline comment or a comment on its own line all depends on the code around you. e.g.
Adding a block comment to a directive at the start of the file always results in the following code:
```
/**
  * flow
 */ 'use strict';
```
To fix this i have created 2 new comment attachment API's, in addition to the existing 2, so you can specify all cases:
- `addLeadingComments` -- Add comments on the line before a specified node.
- `addLeadingInlineComments` -- Add comments inline before a specified node.
- `addTrailingComments` -- Add comments on the line after a specified node.
- `addTrailingInlineComments` -- Add comments inline after a specified node.

In order to achieve this logic i had to reverse engineer prettier's comment placement logic so it will be a fragile with prettier upgrades but all the cases are well tested.

Reviewed By: evanyeung

Differential Revision: D35157487

fbshipit-source-id: 05da53ed745a6ce0ee760290b3ba9ca009af778d