Fix ESTree AST FunctionExpression loc
Summary:
Hermes parser outputs the loc of a property method function to contain the object key in addition to the function. e.g.
Hermes parser function range:

{F714836845}

Flow (and all other parsers):

{F714837077}

This diff patches our ESTree output, via the ESTreeAdapter, to match Flow and all other parsers. The correctness here is somewhat debatable since the `async` property is on the function node but this is how all other parsers do it and we now have tooling that relies on it.

Reviewed By: kassens

Differential Revision: D35162404

fbshipit-source-id: 1268b0d0eafecaefa3a0bb70e6d5a6918bb3a314