Always optimise in CompileJS
Summary:
D20811492 set up a dynamic callback for enabling optimisations that
allowed the caller to decide whether the optimiser should be linked in.
With this change, we can now always have optimisations available in
compileJS, and have the preprocessor flag only control whether
optimisations are available in eval.

Note that we still always link against the optimiser (both before and
after this diff), so we are relying on the linker to drop the symbols
that are not being used.

Credit to mhorowitz for suggesting this as an avenue for exploration.

Reviewed By: kodafb

Differential Revision: D35158309

fbshipit-source-id: 0467f2a541beebdaf9b4aedc4dec00075b726f6f