Merge switchcases in abstract equality
Summary:
A previous change showed that the current organisation of the abstract
equality test can be prone to mistakes, since the first switchcase
performs comparison on HermesValue Tags, which can actually correspond
to multiple different underlying types. This change eliminates the
first switchcase, in favour of always using the second ETag based
switchcase.

It also adds spec annotations, which line up well with this change.

Reviewed By: jpporto

Differential Revision: D35099870

fbshipit-source-id: d4953a048ff58ce4a23d1477e3e3a47d3d4f4384