React faster to external memory buildup
Summary:
Lower the thresholds for triggering an immediate GC on external memory
buildup. This brings the behaviour closer to what would happen if we
allocated memory on the GC heap of that size.

These limits were initially set higher because external memory does not
contribute to pause times. However, we have now seen cases where
infrequent collections in the presence of high external memory allocation
rates result in OOMs.

Reviewed By: kodafb

Differential Revision: D34265810

fbshipit-source-id: 19a030607e9153a09424850ac5a4e6e20f36b5d2