Correctly set gc sections
Summary:
Our check for whether the current build supports gc sections is broken,
and seems to always fail. This diff fixes the check so it passes on
supported builds.

I was under the impression that the Android toolchain also added this
flag, but it appears to only do it for executables, and not for shared
libraries.

We probably want a better system for making sure things like this can't
regress in the open source build, but for now, fix the immediate bug.

Reviewed By: jpporto

Differential Revision: D35273840

fbshipit-source-id: c4ee3c29d760673318b0117d490fe40bf9b133f0