Prevent JS access to internal properties used by structured errors
Summary:
In the current implementation of structured errors, a proxy passed in
as the target of `captureStackTrace` or the target of the `stack`
getter would be able to observe the internal property slot accesses
(and similarly for CallSite). This diff changes:

1. `captureStackTrace` now throws when called with a `Proxy` (this is
the same as V8's behaviour)
2. Use `getOwnNamedDescriptor` instead of `getNamed_RJS`. The former
does not invoke `Proxy` traps.

Reviewed By: jpporto

Differential Revision: D35170292

fbshipit-source-id: 5dfc982b816eea4d98b43aec15b8ef6a064ba2ed