Migrate ICU setup to CMake
Summary:
This stack removes configure.py in favour of directly configuring
CMake. This is done because the current script mostly directly passes
arguments through into CMake anyway, and doesn't support a number of
other possible arguments, making it necessary to supply CMake arguments
as well.

This will make it easier to configure a more optimised build in
upcoming diffs.

To start, this diff migrates the ICU root path setup for devservers to CMake.

Reviewed By: jpporto

Differential Revision: D35342502

fbshipit-source-id: b9634ecb4f701d168ea3264f35f332e744705118