Add build flags for contiguous heap
Summary:
Add Buck and CMake flags to configure whether to use the contiguous
heap. It is turned off when compressed pointers are off since the
advantage of being contiguous is lost in that case. It is also turned
off on Apple platforms due to restrictions on iOS, and on Windows since
we don't currently have the ability to reserve uncommitted memory on
Windows. (although that can be added after this stack)

Reviewed By: kodafb

Differential Revision: D32422495

fbshipit-source-id: 889715e9689cb4d072d2989f006596c28c687768