Use linked list for node list storage
Summary:
Remove the `Vec`s from the AST by using a simple linked list for
`NodeList` instead. At the moment, there's no reason for a doubly linked
list due to the immutability of the AST, so it's only singly linked but
can be changed in the future.

A non-empty `NodeList` is created via `from_iter`, allowing flexibility
for the caller in creation. `NodeList::from_iter` is the only function
that calls `Context::append_list_element`, which is the only means
of modifying a `NodeList`. Thus, `NodeList`s are immutable after
creation, from the perspective of the user.

Storage for the links is separate from the main node storage, to avoid
extra overhead as most nodes won't be elements of a `NodeList`.

Introduce a `mark_lists` function on `Node` in order to mark the
actual list elements. This is necessary because the existing visitor
pattern only visits `Node`s, so we need a way to visit `NodeListElement`
to set the mark bits. `mark_lists` is non-recursive, so it's just a
helper function to mark all the elements.

NOTE: Not all locations are converted to use the linked lists yet:
in particular, the mutable visitor uses `Vec`s to communicate expansion.
Still have to figure out a good way to handle that.

Reviewed By: tmikov

Differential Revision: D34585094

fbshipit-source-id: a7492934953a415c40dd7eea9b150ab462cd3ad6