Create the module scope as a function
Summary:
Currently, we do not correctly process declarations at the top level of
a module because the DeclCollector is only run on functions. This is a
problem for modules, because they are defined as a scope in a dummy
global function, which does not actually contain the AST for the
modules. So when the DeclCollector runs, the declarations in the module
scope cannot be reached.

Using a function also ensures that vars declared at the top level of a
module are not treated as global properties.

Reviewed By: tmikov

Differential Revision: D36231405

fbshipit-source-id: 20a16daa905fe84e659630dc3410d9245ecb1831