add support jsx pragma comment analysis
Summary:
This diff:
1) adds logic to the analyser to consume the jsx pragmas from the docblock
2) adds logic to the analyser to understand fbt

I noticed that we were relying on the `react/jsx-uses-react` lint rule.
For 99% of the code this rule is unnecessary with the hermes scope analyser because it natively supports referencing `React` from JSX.

However there is some code which uses the `Huxpro` pragma to change the component builder to be something different than `React.createElement`!

Reviewed By: pieterv

Differential Revision: D36798240

fbshipit-source-id: 329ad610202b6786976966e85b96f2a000a5a784