Simplify SMI check in SmallHermesValue
Summary:
There are currently two checks when we're determining whether a SMI
will fit inline in a SmallHermesValue:
1. The value remains unchanged after a roundtrip through an int32
2. The int32 can be represented as an int of kNumSmiBits

We can fold away the second check by instead performing a roundtrip
through an int of kNumSmiBits. Such a type doesn't actually exist, but
we can get the same behaviour by sign extending the lower kNumSmiBits
of the int32.

Reviewed By: kodafb

Differential Revision: D35144554

fbshipit-source-id: 47a7fdab2906aa70c2b89cd938ecc68c2396243d