Move code out of juno_ast lib.rs
Summary:
`lib.rs` was getting to have too much code in it and the public API
was getting muddled by the excess of real code we had to read through.

Clean things up by making new files:
* `node_child.rs` contains the actual `NodeChild` trait and all the
  basic node children: lists, nodes, strings, etc.
* `node_enums.rs` contains the trivial enums that we use for things like
  operator kinds, import/export kinds, etc.
* `visitor.rs` contains the `Visitor` patterns and their helpers.

`lib.rs` now more clearly designates the public API for the `juno_ast`
crate and we should continue to ensure that it contains no code.

All the movement is invisible to any users of the crate, as the public
API is completely unchanged.

Reviewed By: tmikov

Differential Revision: D37698636

fbshipit-source-id: 1f6f0800169ff3a466d607286b3b8baa427ee432