improve ast types
Summary:
Adds an explicit parent to a number of nodes. Previously we just used the general `ESNode` parent for everything - however there are a number of cases where we know for certain what the parent **must** be.

For example a `VariableDeclarator` will always be the child of a `VariableDeclaration`.

This makes it a little easier to use the typed AST without adding as many obviously redundant checks.

Reviewed By: pieterv

Differential Revision: D37697813

fbshipit-source-id: 820a8fa0ee4deffd28de2478f374765a0562b536