Use a linked list of freelist heads
Summary:
Hades maintains freelists for each segment in the old generation. When
searching for a cell, it follows these broad steps:
1. Find the smallest size bucket with available cells that can satisfy
the requested allocation.
2. Find a segment that has available cells for that bucket.
3. Traverse the freelist for that size bucket in that segment to get a
cell for the allocation.

Currently, we perform step 2 having an `llvh::SparseBitVector` for each
bucket, and setting a bit in it for each segment. However, lookups in
this data structure can be slow, and adding/removing bits can lead to a
lot of malloc/free activity.

This diff reworks that step to instead introduce a "segment-level
freelist". For each bucket, we now maintain a linked list of the
segments that have available cells for that bucket. With this set up,
we can also avoid the indirection going from the segment index to the
head of its cell freelist, by directly embedding the head in each node
of the segment-level freelist.

This leads to considerably faster old generation allocation times,
which in turn reduces pause times by speeding up young generation
collections.

Reviewed By: jpporto

Differential Revision: D36351191

fbshipit-source-id: 9b672e4bf69a69c125b4e5eec23c261673cf111f