Fix NumberFormat locale lookup
Summary:
There are currently two issues with how we perform locale resolution in
NumberFormat.

1. Instead of passing `resolveLocale` a list of valid locales, we pass
it the locale that the user passed in. This means that it will consider
all locales valid, and does not strip unicode extensions properly,
causing them to be duplicated.
2. We construct the `NSLocale` for the `NSNumberFormatter` using the
locale containing extensions, instead of the base locale. This means
that if the formatter does not recognise some extension, we will get
back an empty string.

Reviewed By: jpporto

Differential Revision: D37887207

fbshipit-source-id: 425680c06e3fea4d1b2a0db7baa952163935f66a