Make BigInt.prototype an ordinary object
Summary:
According to the spec, the BigInt prototype object:
https://tc39.es/ecma262/#sec-properties-of-the-bigint-prototype-object

> is not a BigInt object; it does not have a [[BigIntData]] internal slot

Make it a regular object, and remove the manual check for the prototype
in `thisBigIntValue`.

Note that with this change, we no longer use `fromSignedNoThrow`
outside tests. I'll leave it to jpporto whether we should get rid of
that method entirely, since we can just assert that there isn't a throw
when we're passing it a constant value.

Reviewed By: jpporto

Differential Revision: D37978409

fbshipit-source-id: 9886bccef1642942d2b90fbf0cc487c3e345d891