Remove SynthTraceSerializationTest depending on alloc counts
Summary:
This test fails because the new method for mocking out
non-deterministic functions in synth traces results in a slightly
different number of allocations during recording and replay.

This test also doesn't work under Hades, because Hades doesn't report
the number of allocated objects at the end of each collection, and
because Hades is non-deterministic.

It looks like we weren't running this test because we don't currently
have a test running MallocGC without Handle-SAN.

Differential Revision: D38075261

fbshipit-source-id: 4ea5fffa4a6843ebe15a12a130e7aa6086490f97