mark function variables as used if they have multiple defs and the declaration is exported
Summary:
It's not uncommon to define an overload like this:
```
declare function foo(): void;
declare function foo(arg: string): string;
export function foo(arg?: string) {
  return arg;
}
```

Unfortunately the `no-unused-vars` lint rule is designed such that it only considers the first definition when determining if something is exported.
To work around this limitation, we mark the variable as used if the declaration is exported.

Reviewed By: pieterv

Differential Revision: D38064460

fbshipit-source-id: 7a6a97936b1e743f54806692f10c6550943426be