fix incorrect reference created by a JSX namespace name
Summary:
Namespace names created references to variables when they shouldn't.
This mostly went unnoticed because it's rare that the names clash and lead to a false-negative unused report.
In our flow files we don't use the `no-undef` rule, so we didn't notice it that way until I looked at a non-flow file (which has that rule turned on).

Reviewed By: pieterv

Differential Revision: D38065553

fbshipit-source-id: 9df43906469223d6558058b3842b30e43920ee42