prevent docblock directives from being able to crash the parser
Summary:
I found a hilarious bug where the parser crashed when parsing
```
/**
 * constructor
 */
```

The docblock parser code attempted to call `directives['constructor'].push()` - which crashes because `directives['constructor']` is a function due to the prototypal inheritance of an empty object literal.

`Object.create(null)` creates an object with no prototype - which means we don't have any existing properties to cause weirdness.

Reviewed By: pieterv

Differential Revision: D38087228

fbshipit-source-id: 68710761ab38825f7017b5bb3fecfb1151d6ed82