Use an union inside FNValue
Summary:
The union makes it easier to debug Flow Native with a normal C debugger.

This doesn't introduce new UB because, regardless of whether union type
punning is supported or not, reading an union member different
from the one which was written to and using the value is already UB.

The only valid case is getBool() and that uses memcpy().

Reviewed By: neildhar

Differential Revision: D37266753

fbshipit-source-id: 27cef6610e59fa1b293503fc082ed6c870a7e4fa