Remove ShouldRandomizeAllocSpace
Summary:
This is an unused debug only config that isn't honoured by either GC.
It was originally added to demonstrate bugs when eliding write barriers
in newly created objects. We could support it in Hades, but it wouldn't
serve much purpose, because Hades will always allocate things in the
young gen unless explicitly told not to.

Reviewed By: jpporto

Differential Revision: D38306384

fbshipit-source-id: 75c44603a23f3eaeb4a5d47fa68da7e5f2f89ea8