Move OOM exception into Public
Summary:
Make the OOM exception part of our public API. Our current API rethrows
OOMs as JSI exceptions, however, that makes it impossible to
distinguish recoverable and unrecoverable exceptions. This leads to a
real bug in practice, where if a `HostFunction` results in an OOM, the
JSI translation layer in hermes.cpp will covert it to a regular JS
exception, which can then be caught and discarded.

Later diffs fix the bug, by intentionally not converting OOM exceptions
to JS exceptions.

Reviewed By: jpporto

Differential Revision: D38295849

fbshipit-source-id: 115bb88dbb25bad3ee67e195cd5b296e2e1245b3