Allow pointing ArrayBuffer to an external buffer
Summary:
Add the ability for a JSArrayBuffer to point to an externally provided
data block. When the buffer is marked as pointing to an external data
block, it does not attempt to free it during finalisation, instead the
`NativeState` cell managing the lifetime of the underlying buffer
becomes unreachable, and destroys the external buffer.

This adds no memory cost in the common case, since it only adds a
single bool `external_`, which fits in the padding of the existing
class.

Reviewed By: jpporto

Differential Revision: D37744466

fbshipit-source-id: 747a3f4e5b96f4615cca8c6b6135f5cd29ce56a8