add the shermes tool
Summary:
`shermes` replaces `hermesc` as the compiler driver for Static Hermes.
It is simpler, since it doesn't support all options of `hermesc`, and
its options are closer to a conventional static compiler (`-c`, `-S`,
etc).

The internal configurability of the C compiler invocation is minimal
for now, only supports MacOS. It can be extended easily when needed.

The C include and library paths for invoking the C compiler are derived
from the current build, but they can be overridden by setting
`SHERMES_CC_INCLUDE_PATH` and `SHERMES_CC_LIB_PATH`.

The environment variables `CC`, `CFLAGS`, `LDFLAGS` and `LDLIBS` are
honoured and provide a way to override the above paths.

The `-v` command line option can be used to see the invocation of the C
compiler.

The `-exec` command line option compiles into a temporary binary and
executes it.

Reviewed By: avp

Differential Revision: D39611204

fbshipit-source-id: 1f9460f4fb511e9198d5387630b8e64ca7e15e2e