Add Unresolver pass to resolver (#834)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/834

Support marking identifiers which can be affected by `eval` and `with`
as "unresolvable", much like we do in Juno.

We don't yet support the `canRename` flag, because static Hermes
is as yet unconcerned with outputting JS source,
but there is a comment in the code that explains where to populate such
a flag if it exists.

Reviewed By: tmikov

Differential Revision: D40068199

fbshipit-source-id: b7a3b45bcd18673a3979e3762ede70e88b7ef074