JSON.parse scanString optimization
Summary:
Original Author: fbmal7
Original Git: ba038758418b285d7037d64d7f7a5be1b1b06702

This is the 2nd try at this optimization, [40de6b](https://github.com/facebook/hermes/commit/cc8e1190db858db4fc68596fe88759285e85c513) being the first. If
the underlying stream is utf8 input, the optimization broke, since it
wasn't safe to take the position in the stream. This is because the
same memory was being reused as a conversion buffer.

This new approach adds support in UTF16Stream for "capturing" its
stream. When `beginCapture` is called, the stream will guarantee that
an ArrayRef can be given on the next matching call to `endCapture`,
which will contain all the values the stream produced since
`beginCapture`. If the underlying data is an externally backed `char16_t`
array, then the problem of capturing is trivial. However, in the case
of utf8 input, the Stream will employ two methods to make sure the
active capture it is building is contiguous. These methods are only
needed if it exhausts the current conversion buffer but there is still
an ongoing capture. First, it checks where the beginning of the capture
is. If it is in the second half of the conversion buffer, it simply
moves all of the capture to the beginning of the conversion buffer, and
then resumes from the updated end of the capture. If it's past the
midpoint, it doubles the size of the conversion buffer and updates all
pointers to point to the new storage.

Original Reviewed By: tmikov

Original Revision: D40479705

Reviewed By: tmikov

Differential Revision: D41240823

fbshipit-source-id: cc37079344c687413bbed7447594f7403f9c6310