Improve JS stack profiling
Summary:
Original Author: xidachen@fb.com
Original Git: bda09f5c2b60e5d7055a0c054626496b41ea4880

This is a follow-up to the previous diff(D37889426).

Diff 1: we change the plumbing to pass a pointer to the SamplingProfiler in the callback, so that when we collect the stack trace we know which profiler to collect from.

Diff 2: code clean up to make more code shared between Android and Apple platforms, for the collectStackForLoom() function.

Diff 3: Notify loom when the SamplingProfiler is destroyed, such that it should no longer use the dangling pointer.

Diff 4: Move the enableForLoomCollection() to the point when loom starts, and move the disableForLoomCollection() to the point when loom ends. Add APIs only, these APIs are not used yet in this diff.

Diff5: Split SamplingProfiler::GlobalProfiler::sampleStacks() into two function, the added function does the work of one iteration in sampleStacks().

Diff6: Use the newly added APIs in Diff 4.

Original Revision: D38744022

Reviewed By: neildhar

Differential Revision: D41240806

fbshipit-source-id: ea960fde279f5b95c8926c418115702652b41f13