Split DeleteProperty bytecode into strict and loose variants
Summary:
Create separate strict and loose variants for `DelById` and `DelByVal`.
A subsequent diff will propagate this distinction to the IR.

Also update GetById to use flags from loose mode, instead of
using `defaultPropOpFlags`, since it doesn't actually matter for
getting properties.

With this, we can eliminate the `defaultPropOpFlags` variable in the
Interpreter, since it can now be populated by individual instructions.

Reviewed By: tmikov

Differential Revision: D41297903

fbshipit-source-id: c4915d65feb253f2ba493c70fc9b612dc59117fd