Simplify innerResult handling in iteratorClose
Summary:
Make some simplifications to iteratorClose:
1. Delete some comments that no longer make sense.
2. Check returnRes for undefined instead of checking whether it is a
callable. This is more efficient, and lines up with the spec.
3. Explicitly track whether innerResult is an exception, instead of
relying on thrownValue_ being empty. This matches the rest of our code
and makes it clear that there are no weird side effects going on.
4. Stop holding a mutable handle to the returned object. Instead, just
track whether it was an object in a bool.

Reviewed By: avp

Differential Revision: D41791113

fbshipit-source-id: 752904f92bd21b499f18dfb040dee4e588eceaa6